chore: Move stories to dedicated folder
