+++ web/src/stories/ArtistPresentation.stories.tsx
@@ -6,7 +6,7 @@ import {
   TanStackQueryDecorator,
 } from "@/lib/decorators.tsx";
 
-import { ArtistPresentation } from "./ArtistPresentation";
+import { ArtistPresentation } from "../components/ArtistPresentation.tsx";
 
 const meta = {
   title: "Rallyist/ArtistPresentation",
