+++ web/src/stories/Loader.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import Loader from "./Loader";
+import Loader from "../components/Loader.tsx";
 
 const meta = {
   title: "Atom/Loader",
