+++ web/src/stories/Navbar.RallyistNavbar.stories.tsx
@@ -5,7 +5,7 @@ import {
   RouterDecorator,
 } from "@/lib/decorators.tsx";
 
-import { RallyistNavbar } from "./Navbar";
+import { RallyistNavbar } from "../components/Navbar.tsx";
 
 const meta = {
   title: "Common/Navbar/RallyistNavbar",
