+++ web/src/stories/Navbar.StaffNavbar.stories.tsx
@@ -2,16 +2,18 @@ import type { Meta, StoryObj } from "@storybook/react";
 
 import { RouterDecorator } from "@/lib/decorators.tsx";
 
-import Rules from "./Rules.tsx";
+import { StaffNavbar } from "../components/Navbar.tsx";
 
 const meta = {
-  title: "Routes/Rallyists/Rules",
-  component: Rules,
+  title: "Common/Navbar/StaffNavbar",
+  component: StaffNavbar,
   decorators: [RouterDecorator],
-} satisfies Meta<typeof Rules>;
+} satisfies Meta<typeof StaffNavbar>;
 
 export default meta;
 
 type Story = StoryObj<typeof meta>;
 
-export const Default: Story = {};
+export const Default: Story = {
+  args: {},
+};
