+++ web/src/stories/StandistStaffHomePage.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react";
 
 import { I18nextDecorator, RouterDecorator } from "@/lib/decorators.tsx";
 
-import StandistStaffHomePage from "./StandistStaffHomePage";
+import StandistStaffHomePage from "../components/StandistStaffHomePage.tsx";
 
 const meta = {
   title: "Common/StandistStaffHomePage",
